@@ -4,7 +4,7 @@ version: 0.4.0
 homepage: https://github.com/VolodymyrLykhonis/image_crop
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.4.3 <4.0.0"
   flutter: ">=1.12.13+hotfix.6"
 
 dependencies:
